+++ js/farmOS.map.behaviors.rothamsted_snapshot.js
@@ -1,0 +1,7 @@
+(function () {
+  farmOS.map.behaviors.rothamsted_snapshot = {
+    attach: function (instance) {
+      instance.addBehavior("snapshot");
+    },
+  };
+}(drupalSettings));